# internal/http/swagger/swagger.yaml
basePath: /api/v1
definitions:
  account.UserSettings:
    properties:
      locale:
        type: string
      timezone:
        type: string
    type: object
...
  httpx.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpx.ErrorBody'
      meta: {}
      request_id:
        type: string
      success:
        type: boolean
    type: object
  httpx.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  users.changePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
  users.updateProfileRequest:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      settings:
        $ref: '#/definitions/account.UserSettings'
    type: object
info:
  contact: {}
  description: MrChat backend API for auth, user settings, chat completions, billing,
    and admin configuration. Swagger UI is exposed at /swagger/index.html.
  title: MrChat API
  version: "0.1"
paths:
...
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: Profile payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Users
  /users/me/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Password payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Change current user password
      tags:
      - Users
  /users/me/quota:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Get current user quota
      tags:
      - Users
  /users/me/security:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Get current user security info
      tags:
      - Users
  /users/me/usage:
    get:
      parameters:
      - description: Usage range
        enum:
        - 7d
        - 30d
        - month
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Get current user usage
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter access token as `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"